move the `ansi_styled_msg` helper method out of the base view
This method really only applies if you are using ansi stuff in your
view.  There is no need to have this in the base view.  This moves
the method to the Ansi view helper module and updates the default
view to mixin that module as it needs the helper.  This will help
reduce bloat when building non-ansi views.

This is a cleanup to prep for reworking/simplifying the views when
we switch to accumulating test result data.
