+++ test/unit/view_tests.rb
@@ -35,7 +35,7 @@ class Assert::View
     subject{ @view }
 
     should have_readers :config
-    should have_imeths :view, :is_tty?, :ansi_styled_msg
+    should have_imeths :view, :is_tty?
     should have_imeths :before_load, :after_load
     should have_imeths :on_start, :on_finish, :on_interrupt
     should have_imeths :before_test, :after_test, :on_result
@@ -68,34 +68,6 @@ class Assert::View
 
     should "know if it is a tty" do
       assert_equal !!@io.isatty, subject.is_tty?
-    end
-
-    should "know how to build ansi styled messages" do
-      msg = Factory.string
-      result = [:pass, :fail, :error, :skip, :ignore].sample
-
-      Assert.stub(subject, :is_tty?){ false }
-      Assert.stub(subject, :styled){ false }
-      assert_equal msg, subject.ansi_styled_msg(msg, result)
-
-      Assert.stub(subject, :is_tty?){ false }
-      Assert.stub(subject, :styled){ true }
-      assert_equal msg, subject.ansi_styled_msg(msg, result)
-
-      Assert.stub(subject, :is_tty?){ true }
-      Assert.stub(subject, :styled){ false }
-      assert_equal msg, subject.ansi_styled_msg(msg, result)
-
-      Assert.stub(subject, :is_tty?){ true }
-      Assert.stub(subject, :styled){ true }
-      Assert.stub(subject, "#{result}_styles"){ [] }
-      assert_equal msg, subject.ansi_styled_msg(msg, result)
-
-      styles = Factory.integer(3).times.map{ Assert::ViewHelpers::Ansi::CODES.keys.sample }
-      Assert.stub(subject, "#{result}_styles"){ styles }
-      exp_code = Assert::ViewHelpers::Ansi.code_for(*styles)
-      exp = exp_code + msg + Assert::ViewHelpers::Ansi.code_for(:reset)
-      assert_equal exp, subject.ansi_styled_msg(msg, result)
     end
 
   end
